rename robot_lint.py to lint.py based on code review feedback
