+++ cumulusci/tasks/robotframework/__init__.py
@@ -1,4 +1,4 @@
 from cumulusci.tasks.robotframework.robotframework import Robot  # noqa: F401
 from cumulusci.tasks.robotframework.robotframework import RobotTestDoc  # noqa: F401
 from cumulusci.tasks.robotframework.libdoc import RobotLibDoc  # noqa: F401
-from cumulusci.tasks.robotframework.robot_lint import RobotLint  # noqa: F401
+from cumulusci.tasks.robotframework.lint import RobotLint  # noqa: F401
